api/requirements.txt
fastapi>=0.116,<1.0
python-multipart>=0.0.20,<1.0
pydantic-settings>=2.10,<3.0
PyMuPDF>=1.26,<2.0
python-docx>=1.2,<2.0
transformers>=4.57,<5.0
torch>=2.7,<3.0
uvicorn[standard]>=0.35,<1.0
